New udate

+++ admin/includes/view_all_posts.php
@@ -24,6 +24,28 @@ if (isset($_POST['checkBoxArray'])) {
                     die("QUERY ERROR" . mysqli_error($connection));
                 }
                 break;
+            case 'clone':
+                $query = "SELECT * FROM posts WHERE post_id = '{$postValueID}'";
+                $select_post_query = mysqli_query($connection, $query);
+
+                while($row = mysqli_fetch_array($select_post_query)){
+                    $post_title = $row['post_title'];
+                    $post_category_id = $row['post_category_id'];
+                    $post_date = $row['post_date'];
+                    $post_author = $row['post_author'];
+                    $post_status = $row['post_status'];
+                    $post_image = $row['post_image'];
+                    $post_tags = $row['post_tags'];
+                    $post_content = $row['post_content'];
+                }
+                $query = " INSERT INTO posts(post_category_id, post_title, post_author, post_date, post_status, post_image, post_tags, post_content) ";
+                $query .= "VALUES($post_category_id, '$post_title','$post_author', 
+                '$post_date', '$post_status', '$post_image', '$post_tags', '$post_content')";
+                $copy_query = mysqli_query($connection, $query);
+                if(!$copy_query){
+                    die("Query Failed".mysqli_error($connection));
+                }
+                break;
         }
     }
 }
@@ -39,11 +61,12 @@ if (isset($_POST['checkBoxArray'])) {
                 <option value="published">Publish</option>
                 <option value="draft">Draft</option>
                 <option value="delete">Delete</option>
+                <option value="clone">Clone</option>
             </select>
         </div>
         <div class="col-xs-4" style="margin-bottom: 5px;">
             <input type="submit" name="submit" class="btn btn-success" value="Apply">
-            <a class="btn btn-primary" href="add_post.php">Add New</a>
+            <a class="btn btn-primary" href="posts.php?source=add_post">Add New</a>
         </div>
         <thead class="thead-dark">
             <tr>
@@ -59,14 +82,15 @@ if (isset($_POST['checkBoxArray'])) {
                 <th>Date</th>
                 <th></th>
                 <th></th>
+                <th></th>
             </tr>
         </thead>
         <tbody>
 
             <?php
-            $query = "SELECT * FROM posts";
+            $query = "SELECT * FROM posts ORDER BY post_id DESC";
             $select_posts_query = mysqli_query($connection, $query);
-            while ($row = mysqli_fetch_assoc($select_posts_query)) {
+            while ($row = mysqli_fetch_array($select_posts_query)) {
 
                 $post_id = $row['post_id'];
                 $post_author = $row['post_author'];
@@ -107,8 +131,9 @@ if (isset($_POST['checkBoxArray'])) {
                 echo "<td> $post_tags </td>";
                 echo "<td> $post_comment_count </td>";
                 echo "<td> $post_date </td>";
+                echo "<td> <a class='btn btn-primary' href='../post.php?p_id={$post_id}'> View Post </a> </td>";
                 echo "<td> <a class='btn btn-primary' href='posts.php?source=edit_post&p_id={$post_id}'> Edit </a> </td>";
-                echo "<td> <a class='btn btn-danger' href='posts.php?delete={$post_id}'> Delete </a>​​</td>";
+                echo "<td> <a onClick=\"javascript: return confirm('Are you sure you want to delete!!') \" class='btn btn-danger' href='posts.php?delete={$post_id}'> Delete </a>​​</td>";
                 echo "</tr>";
             }
             ?>
